nlp-module/requirements.txt
# Web framework
Flask==3.0.0
flask-cors==3.0.10

# GraphQL client
gql==3.4.1
graphql-core==3.2.3

# HTTP client for GraphQL
aiohttp==3.9.1

# NLP libraries
spacy==3.7.2
nltk==3.8.1

# OpenAI and its dependencies
openai==1.3.5
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Additional dependencies for spaCy and NLTK
packaging==23.2
typing-extensions==4.9.0

# RDFLib
rdflib==6.0.2